Add files via upload
Voltage Doubler DC Boost Converter



Hello everyone,

In this project, I designed a DC boost converter and wanted to share it with you. It is quite simple to design the circuit using only 2x1N4007 diodes and 2x1000uF electrolyte capacitors as components. I also added 4 screw holes for fixing it to the area you will use.

Maximum voltages of 6,12,18,24 or 50V can be given as input voltage. At the output, you can convert voltage up to 12,24,36,48 or 100V. You can order a circuit that is very simple to assemble and use it easily. I have attached the required documents in the description section.